fix(stark): use AtomicUsize for gpu_layer_circuit_id (#241 prereq)
The static was AtomicU64 — fine on 64-bit host but fails to resolve on
the mipsel-zkm-zkvm-elf target because that arch lacks
target_has_atomic="64". The whole zkvm-elf build of zkm-stark
(triggered transitively by test-artifacts) was broken.

SP1 follows the same pattern: AtomicUsize on cross-target paths
(slop/jit/shm.rs) and reserves AtomicU64 for host-only worker code
that doesn't compile to zkvm. Public u64 API preserved via cast.

feat(recursion): #241 Phase 1 — additive jagged-PCS Witnessable impls

Lays the witness-stream foundation for porting SP1's jagged-PCS proof
witnessing to Ziren. Pure addition — no field-type changes, no call-site
rewires. Phase 2-4 (eval→coeff Lagrange + bundle assembly + lift
deletion) follow per task #241.

Adds Witnessable<C> for:
  * JaggedReductionRound<InnerChallenge>     (3 EF evals)
  * JaggedReductionProof<InnerChallenge>     (Vec<round> + eval_point + q_at_z)
  * LeafOpening<InnerVal, LbMmcs>            (Vec<Vec<F>> + MT::Proof const)
  * MerkleOpening<InnerVal, LbMmcs>          (Vec<LeafOpening>)

Plus companion in-circuit variable types (LeafOpeningVar, MerkleOpeningVar)
and 4 smoke tests covering the field-by-field witness mapping.

+++ crates/stark/src/basefold_late_binding.rs
@@ -703,8 +703,16 @@ pub fn get_gpu_layer_drain_circuit_hook() -> Option<GpuLayerDrainCircuitFn> {
 /// `(device_id, circuit_id)` so concurrent shards on the same GPU are
 /// fully isolated — fixes #230 multi-GPU panics caused by a shared
 /// `next_handle` counter being stepped on across shards.
-static NEXT_GPU_LAYER_CIRCUIT_ID: std::sync::atomic::AtomicU64 =
-    std::sync::atomic::AtomicU64::new(1);
+// Backing storage uses AtomicUsize, not AtomicU64, so the file
+// compiles on the zkvm-elf target (mipsel — no
+// `target_has_atomic="64"`).  The GPU registry never executes on the
+// zkvm-elf binary, but the symbol still has to type-check in that
+// build because `row_gkr/build.rs` imports the helper unconditionally.
+// Public API (`u64`) is preserved via cast.  On host (64-bit)
+// `usize == u64`; on the 32-bit zkvm-elf the upper bits are always
+// zero and circuit IDs grow well within `u32::MAX`.
+static NEXT_GPU_LAYER_CIRCUIT_ID: std::sync::atomic::AtomicUsize =
+    std::sync::atomic::AtomicUsize::new(1);
 
 /// Allocate a fresh process-unique GKR-circuit ID for use with the
 /// GPU layer-state hooks.  Must be called once per
@@ -718,7 +726,7 @@ static NEXT_GPU_LAYER_CIRCUIT_ID: std::sync::atomic::AtomicU64 =
 /// outside the threat model.
 #[must_use]
 pub fn allocate_gpu_layer_circuit_id() -> u64 {
-    NEXT_GPU_LAYER_CIRCUIT_ID.fetch_add(1, std::sync::atomic::Ordering::Relaxed)
+    NEXT_GPU_LAYER_CIRCUIT_ID.fetch_add(1, std::sync::atomic::Ordering::Relaxed) as u64
 }
 
 /// Open the committed batch at a single point and produce the
